(MAINT) Restructure project dependencies
Prior to this commit there were additional/incorrect dependencies
included in this project. Given that we only need the bare minimum to
get start with the linter this commit removes anything that is not
nessesary for now.

Additionally the Rakefile has also been cleaned up and includes an
enhanced task for running only spec tests.