(CONT-234) Fix failing parameterised commmand test
Prior to this commit, the MVP failed a few tests. One such test expects a parameterised command (ie the command parameters are placed in an array) to detect 0 problems. The MVP failed this test as it raised a warning if an input variable was found regardless of the context in which it was found.

This commit fixes this failing test by adding a few functions - 1) checks if a command is parameterised 2) puts the logic which iterates over a command and checks for violations into it's own function 3) extracts the logic which raises a warning into its own function.

This commit also reduces some complexity in how each command is parsed. Previously, there were leading tokens included at the start of each command that were unnecessary to check during iterations. This commit checks for the `command` token and starts a while loop from there to iterate over each token in the command. This makes the logic of the code more straight forward and easier to understand.