CONT-234_fix-title-test
Prior to this commit, the title_tokens function provided by puppet lint was only able to successfully parse titles in single quotes. Titles in double quotes are of interest to us as they may contain unsafe interpolated variables.

This commit adds a modified function for title_tokens called get_title_tokens and checks the result for any interpolated variables.
This function returns a hash containing an array of tokens and the resource type.
Another test has been added to test resources with an array of titles.