CONT-234 Refactor nested loops
The check iterates over an array of exec commands and then over each token in each command. Prior to this commit, these loops were nested. The goal of this commit is to refactor this to clean up the code a bit. This was achieved by extracting the logic which iterates over the tokens in a command into a function called check_unsafe_interpolations. Separating code logic into functions and removing nested loops improves code readability and should make it easier to build upon in the future.